♻️ refactor: useActivityReviews.ts 커스텀 훅 생성 리팩토링을 통해 커스텀 훅을 생성했습니다. Resolve: #143
